+++ BackroomsRave/MovingLamp.h
@@ -7,6 +7,10 @@
 #include <random>
 #include "types.h"
 
+/*
+ * A lamp that moves back and forth between two points, 
+ * changing color randomly.
+ */
 class MovingLamp {
 public:
 	MovingLamp(DirectX::XMFLOAT3 start, DirectX::XMFLOAT3 end, float speed);
@@ -26,7 +30,8 @@ private:
 	float t;
 	bool forward;
 	std::chrono::time_point<std::chrono::steady_clock> last_color_change;
-	const uint64_t color_change_interval_ms = 60000.f / 165.0f; // 165 BPM, same as the music
+	const int64_t color_change_interval_ms = 
+		static_cast<int64_t>(60000.f / 165.0f); // 165 BPM, same as the music
 	std::vector<square_instance_t> instances;
 	std::mt19937 gen;
 };
